requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
google-generativeai==0.4.0
xgboost==2.0.3
prophet==1.1.5
shap==0.44.1
openpyxl==3.1.2
python-dotenv==1.0.1
